feat: add header component
header comes with Title, Search bar, City Dropdown and UserProfile. Also included user and theme contexts.
This is just a overall basic layout. Each of them individually has to be
updated. Faced issues with aligning the items in the header. Especially
the title and search bar. Need to raise a issue for this and try to fix
later.